refactor(kaggle): 优化Kaggle T4 GPU配置和简化代码
- 统一使用FP32数据类型确保兼容性
- 简化GPU缓存和预处理逻辑
- 移除冗余代码和测试文件
- 调整批次大小和worker数量以适应T4 GPU
- 更新文档和注释以反映优化配置

+++ kaggle_cells/cell_03_data_processing.py
@@ -43,8 +43,8 @@ def extract_frames_gpu_accelerated(video_path, max_frames=16, target_size=(224, 
         
         # GPU上进行质量检测（使用Sobel算子代替Laplacian）
         if quality_threshold > 0:
-            # 转换为灰度图进行质量检测
-            gray_frames = selected_frames.mean(dim=-1, keepdim=True)  # (T, H, W, 1)
+            # 转换为灰度图进行质量检测（先转换为float类型）
+            gray_frames = selected_frames.float().mean(dim=-1, keepdim=True)  # (T, H, W, 1)
             gray_frames = gray_frames.permute(0, 3, 1, 2)  # (T, 1, H, W)
             
             # 使用Sobel算子计算图像质量
